Check make on mac and some norm
