+++ maps/bad/xpm_is_jpg.cub
@@ -1,0 +1,17 @@
+NO textures/fake.xpm
+SO textures/fake.xpm
+WE textures/fake.xpm
+EA textures/fake.xpm
+
+C 0,191,255
+F 60,48,42
+
+1111111111111111111111111
+1000000000000000000100001
+1001000000000S00000000001
+1001000000000000001000001
+1001000000000000001000001
+1001000000100000001000001
+1001000000000000001000001
+1001000000001000001000001
+1111111111111111111111111
